+++ lang/en/messages.php
@@ -9,4 +9,9 @@ return [
     'partially-missing' => 'Partially missing',
     'translated' => 'Translated',
 
+    'email.hello' => 'Hello!',
+    'email.regards' => 'Regards',
+    'email.signature' => 'Team :company',
+    'email.trouble' => 'If you\'re having trouble clicking the button, click the URL below or copy and paste it into your web browser.',
+
 ];