+++ resources/views/components/emails/button-plain.blade.php
@@ -1,0 +1,6 @@
+@props([
+    'link',
+])
+
+-- {{ $slot }}
+{{ $link }}