+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.5.0] - 2025-03-13
+
+_Stable release based on [0.5.0-rc.2]._
+
+## [0.5.0-rc.2] - 2025-03-13
+
+### Added
+
+- Add email layouts, components and translations.
+
+## [0.5.0-rc.1] - 2025-02-19
+
 ## [0.4.7] - 2025-03-04
 
 ### Fixed
@@ -104,6 +116,9 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[0.5.0]: https://https://github.com/internetguru/laravel-common/compare/v0.4.7...v0.5.0
+[0.5.0-rc.2]: https://github.com/internetguru/laravel-common/releases/tag/v0.4.7
+[0.5.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v0.4.2
 [0.4.7]: https://https://github.com/internetguru/laravel-common/compare/v0.4.6...v0.4.7
 [0.4.6]: https://https://github.com/internetguru/laravel-common/compare/v0.4.5...v0.4.6
 [0.4.5]: https://https://github.com/internetguru/laravel-common/compare/v0.4.4...v0.4.5